+++ CHANGELOG.md
@@ -6,6 +6,7 @@ It is based on Tidepool Blip 1.27.
 
 ### Engineering use
 - YLP-1626 Refactor team hook to only keep functions impacting the state
+- YLP-1631 Install react testing library
 
 ## 3.0.0 - 2022-06-16
 ### Added